+++ src/BlazorUI/Bit.BlazorUI.Tests/SearchBoxes/BitSearchBoxValidationTest.razor
@@ -3,7 +3,6 @@
 
     <BitSearchBox IsEnabled="IsEnabled"
                   Placeholder="@Placeholder"
-                  TItem="BitSearchBoxItem"
                   @bind-Value="@TestModel.Value"
                   DisableAnimation="@DisableAnimation"
                   IsUnderlined="@IsUnderlined"
