+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBox.razor
@@ -1,10 +1,5 @@
 ﻿@namespace Bit.BlazorUI
 @inherits BitInputBase<string?>
-@typeparam TItem
-
-<CascadingValue Value="this" IsFixed="true">
-    <div style="display:none" hidden>@(Options ?? ChildContent)</div>
-</CascadingValue>
 
 <div @ref="RootElement" @attributes="HtmlAttributes"
      id="@_Id"
@@ -15,16 +10,16 @@
             <i style="@Styles?.SearchIcon" class="bit-icon bit-icon--@IconName @Classes?.SearchIcon" aria-hidden="true" />
         </div>
         <input @ref="_inputRef" @attributes="InputHtmlAttributes" @onkeydown="@((k) => HandleOnKeyDown(k))" @onfocusin="HandleInputFocusIn" @onfocusout="HandleInputFocusOut" @oninput="@HandleOnChange"
-                id="@_inputId"
-                style="@Styles?.Input"
-                class="bit-srb-inp @Classes?.Input"
-                type="text"
-                role="searchbox"
-                value="@_searchValue"
-                aria-label="@AriaLabel"
-                placeholder="@Placeholder"
-                autocomplete="@Autocomplete"
-                disabled="@(IsEnabled is false)" />
+               id="@_inputId"
+               style="@Styles?.Input"
+               class="bit-srb-inp @Classes?.Input"
+               type="text"
+               role="searchbox"
+               value="@_searchValue"
+               aria-label="@AriaLabel"
+               placeholder="@Placeholder"
+               autocomplete="@Autocomplete"
+               disabled="@(IsEnabled is false)" />
         @if (CurrentValue.HasValue())
         {
             <div style="@Styles?.ClearButtonContainer" class="bit-srb-cbc @Classes?.ClearButtonContainer">
@@ -43,50 +38,40 @@
         }
     </div>
 
-    <div @onclick="CloseCallout"
-         style="display:@(isOpen ? "block" : "none");@Styles?.Overlay"
-         class="bit-srb-ovl @Classes?.Overlay"></div>
+    @if ((SuggestedItems?.Any() ?? false) || (Virtualize && SuggestedItemsProvider is not null))
+    {
+        <div @onclick="CloseCallout"
+             style="display:@(isOpen ? "block" : "none");@Styles?.Overlay"
+             class="bit-srb-ovl @Classes?.Overlay"></div>
 
-    <div id="@_calloutId"
-         tabindex="0"
-         role="listbox"
-         style="@Styles?.Callout"
-         class="bit-srb-cal @Classes?.Callout">
+        <div id="@_calloutId"
+             tabindex="0"
+             role="listbox"
+             style="@Styles?.Callout"
+             class="bit-srb-cal @Classes?.Callout">
 
-        <div id="@_scrollContainerId"
-             style="@Styles?.ScrollContainer"
-             class="bit-srb-scn @Classes?.ScrollContainer">
-            @if (Virtualize)
-            {
-                if (ItemsProvider is null)
+            <div id="@_scrollContainerId"
+                 style="@Styles?.ScrollContainer"
+                 class="bit-srb-scn @Classes?.ScrollContainer">
+                @if (Virtualize && SuggestedItemsProvider is not null)
                 {
-                    <Virtualize Items="_searchItems"
+                    <Virtualize @ref="_virtualizeElement"
                                 Context="item"
-                                ItemSize="ItemSize"
-                                OverscanCount="OverscanCount"
-                                Placeholder="VirtualizePlaceholder">
-                        <_BitSearchBoxItem Item="item" SearchBox="this" />
+                                ItemSize="SuggestedItemSize"
+                                Placeholder="VirtualizePlaceholder"
+                                ItemsProvider="InternalItemsProvider">
+                        <_BitSearchBoxSuggestedItem Text="@item" SearchBox="this" />
                     </Virtualize>
                 }
                 else
                 {
-                    <Virtualize @ref="_virtualizeElement"
-                                Context="item"
-                                OverscanCount="OverscanCount"
-                                Placeholder="VirtualizePlaceholder"
-                                ItemsProvider="InternalItemsProvider">
-                        <_BitSearchBoxItem Item="item" SearchBox="this" />
-                    </Virtualize>
+                    @foreach (var item in _searchItems)
+                    {
+                        <_BitSearchBoxSuggestedItem Text="@item" SearchBox="this" />
+                    }
                 }
-            }
-            else
-            {
-                @foreach (var item in _searchItems)
-                {
-                    <_BitSearchBoxItem Item="item" SearchBox="this" />
-                }
-            }
-        </div>
+            </div>
 
-    </div>
+        </div>
+    }
 </div>
