--- src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBoxItemsProvider.cs
@@ -1,3 +1,0 @@
-﻿namespace Bit.BlazorUI;
-
-public delegate ValueTask<BitSearchBoxItemsProviderResult<TItem>> BitSearchBoxItemsProvider<TItem>(BitSearchBoxItemsProviderRequest<TItem> request);
