--- src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBoxNameSelectors.cs
@@ -1,39 +1,0 @@
-﻿namespace Bit.BlazorUI;
-
-public class BitSearchBoxNameSelectors<TItem>
-{
-    /// <summary>
-    /// The AriaLabel field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> AriaLabel { get; set; } = new(nameof(BitSearchBoxItem.AriaLabel));
-
-    /// <summary>
-    /// The CSS Class field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> Class { get; set; } = new(nameof(BitSearchBoxItem.Class));
-
-    /// <summary>
-    /// The Id field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> Id { get; set; } = new(nameof(BitSearchBoxItem.Id));
-
-    /// <summary>
-    /// The IsSelected field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, bool> IsSelected { get; set; } = new(nameof(BitSearchBoxItem.IsSelected));
-
-    /// <summary>
-    /// The CSS Style field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> Style { get; set; } = new(nameof(BitSearchBoxItem.Style));
-
-    /// <summary>
-    /// The Text field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> Text { get; set; } = new(nameof(BitSearchBoxItem.Text));
-
-    /// <summary>
-    /// The Title field name and selector of the custom input class.
-    /// </summary>
-    public BitNameSelectorPair<TItem, string?> Title { get; set; } = new(nameof(BitSearchBoxItem.Title));
-}
