+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBoxSuggestedItem.cs
@@ -1,0 +1,13 @@
+﻿//namespace Bit.BlazorUI;
+//internal class BitSearchBoxSuggestedItem
+//{
+//    /// <summary>
+//    /// Determines if the suggested item is selected.
+//    /// </summary>
+//    public bool IsSelected { get; set; }
+
+//    /// <summary>
+//    /// The text to render for the suggested item.
+//    /// </summary>
+//    public string? Text { get; set; }
+//}
