+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/BitSearchBoxSuggestedItemsProvider.cs
@@ -1,0 +1,3 @@
+﻿namespace Bit.BlazorUI;
+
+public delegate ValueTask<BitSearchBoxSuggestedItemsProviderResult<TItem>> BitSearchBoxSuggestedItemsProvider<TItem>(BitSearchBoxSuggestedItemsProviderRequest<TItem> request);
