--- src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/_BitSearchBoxItem.razor
@@ -1,35 +1,0 @@
-@namespace Bit.BlazorUI
-@typeparam TItem
-
-@{
-    var id = SearchBox.GetId(Item);
-    var text = SearchBox.GetText(Item);
-    var isSelected = SearchBox.GetIsSelected(Item);
-}
-
-<div style="@SearchBox.GetStyle(Item)" class="@SearchBox.GetClass(Item)">
-    <button @onclick="()=> SearchBox.HandleOnItemClick(Item)"
-            id="@id"
-            type="button"
-            role="option"
-            title="@SearchBox.GetTitle(Item)"
-            tabindex="0"
-            style="@SearchBox.Styles?.ItemButton"
-            class="bit-srb-itm@(isSelected ? $" bit-srb-sel" : "") @SearchBox.Classes?.ItemButton"
-            aria-label="@SearchBox.GetAriaLabel(Item)"
-            aria-setsize="@SearchBox.GetTotalItems()"
-            aria-selected="@(isSelected ? "true" : "false")"
-            aria-posinset="@SearchBox.GetItemPosInSet(Item)">
-        @if (SearchBox.ItemTemplate is not null)
-        {
-            @SearchBox.ItemTemplate(Item)
-        }
-        else
-        {
-            <span class="@SearchBox.Classes?.ItemText"
-                  style="@SearchBox.Styles?.ItemText">
-                @text
-            </span>
-        }
-    </button>
-</div>
