--- src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/_BitSearchBoxItem.razor.cs
@@ -1,8 +1,0 @@
-﻿namespace Bit.BlazorUI;
-
-public partial class _BitSearchBoxItem<TItem> : ComponentBase where TItem : class
-{
-    [Parameter] public TItem Item { get; set; } = default!;
-
-    [Parameter] public BitSearchBox<TItem> SearchBox { get; set; } = default!;
-}
