+++ src/BlazorUI/Bit.BlazorUI/Components/Inputs/SearchBox/_BitSearchBoxSuggestedItem.razor.cs
@@ -1,0 +1,8 @@
+﻿namespace Bit.BlazorUI;
+
+public partial class _BitSearchBoxSuggestedItem : ComponentBase
+{
+    [Parameter] public string Text { get; set; } = default!;
+
+    [Parameter] public BitSearchBox SearchBox { get; set; } = default!;
+}
