--- src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/SearchBox/BitSearchBoxCustom.cs
@@ -1,17 +1,0 @@
-﻿namespace Bit.BlazorUI.Demo.Client.Core.Pages.Components.Inputs.SearchBox;
-public class BitSearchBoxCustom
-{
-    public string? Label { get; set; }
-
-    public string? CssClass { get; set; }
-
-    public string? CssStyle { get; set; }
-
-    public string? Key { get; set; }
-
-    public string? Text { get; set; }
-
-    public string? Title { get; set; }
-
-    public bool IsSelected { get; set; }
-}
