--- src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/SearchBox/_BitSearchBoxItemDemo.razor
@@ -1,147 +1,0 @@
-﻿<ComponentExampleBox Title="Basic" RazorCode="@example1RazorCode" Id="example1">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Basic</BitLabel>
-                <BitSearchBox Placeholder="Search" TItem="BitSearchBoxItem" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Disabled</BitLabel>
-                <BitSearchBox Placeholder="Search" IsEnabled="false" TItem="BitSearchBoxItem" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="IsUnderlined" RazorCode="@example2RazorCode" Id="example2">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Basic Underlined SearchBox</BitLabel>
-                <BitSearchBox Placeholder="Search" IsUnderlined="true" TItem="BitSearchBoxItem" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Disabled Underlined SearchBox</BitLabel>
-                <BitSearchBox Placeholder="Search" IsUnderlined="true" IsEnabled="false" TItem="BitSearchBoxItem" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Icon" RazorCode="@example3RazorCode" Id="example3">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>SearchBox with fixed icon</BitLabel>
-                <BitSearchBox Placeholder="Search" FixedIcon="true" TItem="BitSearchBoxItem" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>SearchBox without icon animation</BitLabel>
-                <BitSearchBox Placeholder="Search" DisableAnimation="true" TItem="BitSearchBoxItem" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>SearchBox with custom icon</BitLabel>
-                <BitSearchBox Placeholder="Search" IconName="@BitIconName.Filter" TItem="BitSearchBoxItem" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Style & Class" RazorCode="@example4RazorCode" Id="example4">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Component's Style & Class:</BitLabel>
-                <BitSearchBox Placeholder="Search" Style="background-color: lightskyblue; border-radius: 1rem; margin: 1rem 0" TItem="BitSearchBoxItem" />
-                <BitSearchBox Placeholder="Search" Class="custom-class" TItem="BitSearchBoxItem" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Styles & Classes:</BitLabel>
-                <BitSearchBox Placeholder="Search"
-                              Styles="@(new() { SearchIcon = "color: darkorange;",
-                                                Input = "padding: 0.5rem; background-color: goldenrod" })" 
-                              TItem="BitSearchBoxItem" />
-                <br />
-                <BitSearchBox Placeholder="Search" DefaultValue="This is default value"
-                              Classes="@(new() { ClearButtonIcon = "custom-clear",
-                                                 SearchIconContainer = "custom-search" })"
-                              TItem="BitSearchBoxItem" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Visibility" RazorCode="@example5RazorCode" Id="example5">
-    <ExamplePreview>
-        <div class="example-content">
-            <div class="visibility-items">Visible: [ <BitSearchBox Visibility="BitVisibility.Visible" Placeholder="Visible SearchBox" TItem="BitSearchBoxItem" /> ]</div>
-            <div class="visibility-items">Hidden: [ <BitSearchBox Visibility="BitVisibility.Hidden" Placeholder="Hidden SearchBox" TItem="BitSearchBoxItem" />  ]</div>
-            <div class="visibility-items">Collapsed: [ <BitSearchBox Visibility="BitVisibility.Collapsed" Placeholder="Collapsed SearchBox" TItem="BitSearchBoxItem" />  ]</div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Binding" RazorCode="@example6RazorCode" CsharpCode="@example6CsharpCode" Id="example6">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Two-way Bind</BitLabel>
-                <BitSearchBox Placeholder="Search" TItem="BitSearchBoxItem" @bind-Value="TwoWaySearchValue" />
-                <BitTextField Placeholder="Search Value" Style="margin-top: 5px;" @bind-Value="TwoWaySearchValue" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>OnChange</BitLabel>
-                <BitSearchBox Placeholder="Search" OnChange="(s) => OnChangeSearchValue = s" OnClear="() => OnChangeSearchValue = string.Empty" TItem="BitSearchBoxItem" />
-                <BitLabel>Search Value: @OnChangeSearchValue</BitLabel>
-            </div>
-            <br />
-            <div>
-                <BitLabel>OnSearch (Search by Enter)</BitLabel>
-                <BitSearchBox Placeholder="Search" OnSearch="(s) => OnSearchValue = s" OnClear="() => OnSearchValue = string.Empty" TItem="BitSearchBoxItem" />
-                <BitLabel>Search Value: @OnSearchValue</BitLabel>
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Validation" RazorCode="@example7RazorCode" CsharpCode="@example7CsharpCode" Id="example7">
-    <ExamplePreview>
-        <div class="example-content">
-            <EditForm Model="ValidationSearchBoxModel">
-                <DataAnnotationsValidator />
-                <BitSearchBox Placeholder="Search" DefaultValue="This is default value" TItem="BitSearchBoxItem" @bind-Value="ValidationSearchBoxModel.Text" />
-                <ValidationMessage For="() => ValidationSearchBoxModel.Text" />
-            </EditForm>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Suggest List" RazorCode="@example8RazorCode" CsharpCode="@example8CsharpCode" Id="example8">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Items:</BitLabel>
-                <BitSearchBox Placeholder="e.g. Apple" Items="GetBasicItems()" @bind-Value="@SearchValue" />
-                <br />
-                SearchValue: @SearchValue
-            </div>
-
-            <div>
-                <BitLabel>ItemsProvider:</BitLabel>
-                <BitSearchBox Placeholder="e.g. Pro"
-                              Virtualize="true"
-                              ItemsProvider="LoadItems" 
-                              TItem="BitSearchBoxItem" 
-                              @bind-Value="@ItemsProviderSearchValue" />
-                <br />
-                SearchValue: @ItemsProviderSearchValue
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
