--- src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/SearchBox/_BitSearchBoxOptionDemo.razor
@@ -1,145 +1,0 @@
-﻿<ComponentExampleBox Title="Basic" RazorCode="@example1RazorCode" Id="example1">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Basic</BitLabel>
-                <BitSearchBox Placeholder="Search" TItem="BitSearchBoxOption" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Disabled</BitLabel>
-                <BitSearchBox Placeholder="Search" IsEnabled="false" TItem="BitSearchBoxOption" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="IsUnderlined" RazorCode="@example2RazorCode" Id="example2">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Basic Underlined SearchBox</BitLabel>
-                <BitSearchBox Placeholder="Search" IsUnderlined="true" TItem="BitSearchBoxOption" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Disabled Underlined SearchBox</BitLabel>
-                <BitSearchBox Placeholder="Search" IsUnderlined="true" IsEnabled="false" TItem="BitSearchBoxOption" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Icon" RazorCode="@example3RazorCode" Id="example3">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>SearchBox with fixed icon</BitLabel>
-                <BitSearchBox Placeholder="Search" FixedIcon="true" TItem="BitSearchBoxOption" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>SearchBox without icon animation</BitLabel>
-                <BitSearchBox Placeholder="Search" DisableAnimation="true" TItem="BitSearchBoxOption" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>SearchBox with custom icon</BitLabel>
-                <BitSearchBox Placeholder="Search" IconName="@BitIconName.Filter" TItem="BitSearchBoxOption" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Style & Class" RazorCode="@example4RazorCode" Id="example4">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Component's Style & Class:</BitLabel>
-                <BitSearchBox Placeholder="Search" Style="background-color: lightskyblue; border-radius: 1rem; margin: 1rem 0" TItem="BitSearchBoxOption" />
-                <BitSearchBox Placeholder="Search" Class="custom-class" TItem="BitSearchBoxOption" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>Styles & Classes:</BitLabel>
-                <BitSearchBox Placeholder="Search"
-                              Styles="@(new() { SearchIcon = "color: darkorange;",
-                                                Input = "padding: 0.5rem; background-color: goldenrod" })"
-                              TItem="BitSearchBoxOption" />
-                <br />
-                <BitSearchBox Placeholder="Search" DefaultValue="This is default value"
-                              Classes="@(new() { ClearButtonIcon = "custom-clear",
-                                                 SearchIconContainer = "custom-search" })"
-                              TItem="BitSearchBoxOption" />
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Visibility" RazorCode="@example5RazorCode" Id="example5">
-    <ExamplePreview>
-        <div class="example-content">
-            <div class="visibility-items">Visible: [ <BitSearchBox Visibility="BitVisibility.Visible" Placeholder="Visible SearchBox" TItem="BitSearchBoxOption" /> ]</div>
-            <div class="visibility-items">Hidden: [ <BitSearchBox Visibility="BitVisibility.Hidden" Placeholder="Hidden SearchBox" TItem="BitSearchBoxOption" />  ]</div>
-            <div class="visibility-items">Collapsed: [ <BitSearchBox Visibility="BitVisibility.Collapsed" Placeholder="Collapsed SearchBox" TItem="BitSearchBoxOption" />  ]</div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Binding" RazorCode="@example6RazorCode" CsharpCode="@example6CsharpCode" Id="example6">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitLabel>Two-way Bind</BitLabel>
-                <BitSearchBox Placeholder="Search" TItem="BitSearchBoxOption" @bind-Value="TwoWaySearchValue" />
-                <BitTextField Placeholder="Search Value" Style="margin-top: 5px;" @bind-Value="TwoWaySearchValue" />
-            </div>
-            <br />
-            <div>
-                <BitLabel>OnChange</BitLabel>
-                <BitSearchBox Placeholder="Search" OnChange="(s) => OnChangeSearchValue = s" OnClear="() => OnChangeSearchValue = string.Empty" TItem="BitSearchBoxOption" />
-                <BitLabel>Search Value: @OnChangeSearchValue</BitLabel>
-            </div>
-            <br />
-            <div>
-                <BitLabel>OnSearch (Search by Enter)</BitLabel>
-                <BitSearchBox Placeholder="Search" OnSearch="(s) => OnSearchValue = s" OnClear="() => OnSearchValue = string.Empty" TItem="BitSearchBoxOption" />
-                <BitLabel>Search Value: @OnSearchValue</BitLabel>
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Validation" RazorCode="@example7RazorCode" CsharpCode="@example7CsharpCode" Id="example7">
-    <ExamplePreview>
-        <div class="example-content">
-            <EditForm Model="ValidationSearchBoxModel">
-                <DataAnnotationsValidator />
-                <BitSearchBox Placeholder="Search" DefaultValue="This is default value" TItem="BitSearchBoxOption" @bind-Value="ValidationSearchBoxModel.Text" />
-                <ValidationMessage For="() => ValidationSearchBoxModel.Text" />
-            </EditForm>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
-
-<ComponentExampleBox Title="Suggest List" RazorCode="@example8RazorCode" CsharpCode="@example8CsharpCode" Id="example8">
-    <ExamplePreview>
-        <div class="example-content">
-            <div>
-                <BitSearchBox Placeholder="e.g. Apple"
-                              TItem="BitSearchBoxOption"
-                              @bind-Value="@SearchValue">
-                    <BitSearchBoxOption Text="Apple" />
-                    <BitSearchBoxOption Text="Banana" />
-                    <BitSearchBoxOption Text="Orange" />
-                    <BitSearchBoxOption Text="Grape" />
-                    <BitSearchBoxOption Text="Broccoli" />
-                    <BitSearchBoxOption Text="Carrot" />
-                    <BitSearchBoxOption Text="Lettuce" />
-                </BitSearchBox>
-                <br />
-                SearchValue: @SearchValue
-            </div>
-        </div>
-    </ExamplePreview>
-</ComponentExampleBox>
