--- src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Inputs/SearchBox/_BitSearchBoxOptionDemo.razor.cs
@@ -1,11 +1,0 @@
-﻿namespace Bit.BlazorUI.Demo.Client.Core.Pages.Components.Inputs.SearchBox;
-
-public partial class _BitSearchBoxOptionDemo
-{
-    private string TwoWaySearchValue;
-    private string OnChangeSearchValue;
-    private string OnSearchValue;
-    private string SearchValue;
-
-    private readonly ValidationSearchBoxModel ValidationSearchBoxModel = new();
-}
