+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/Components/Panel/BitPanelDemo.razor
@@ -49,7 +49,7 @@
                 <BitPanel Classes="@panelClassStyles" @bind-IsOpen="IsPanelWithCustomHeaderOpen">
                     <HeaderTemplate>
                         <div>
-                            <BitSearchBox Placeholder="Search here..." TItem="BitSearchBoxItem" />
+                            <BitSearchBox Placeholder="Search here..." />
                             <p>
                                 BitPanel with custom header content
                             </p>
