+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Pages/IconographyPage.razor
@@ -23,8 +23,7 @@
                       IconName="@BitIconName.Search"
                       Value="@searchText"
                       ValueChanged="@HandleChange"
-                      OnClear="@HandleClear"
-                      TItem="BitSearchBoxItem" />
+                      OnClear="@HandleClear" />
         <div class="icon-list">
             @if (filteredIcons.Count == 0)
             {
