+++ src/BlazorUI/Demo/Client/Bit.BlazorUI.Demo.Client.Core/Shared/NavMenu.razor
@@ -12,8 +12,7 @@
                       IsUnderlined="true"
                       Value="@_searchText"
                       ValueChanged="@HandleChange"
-                      OnClear="@HandleClear"
-                      TItem="BitSearchBoxItem" />
+                      OnClear="@HandleClear" />
         @if (_filteredNavItems.Count == 0)
         {
             <br />
